Refactor maintenance page and component
Added a hook to handle the maintenance logic too
